requirements.txt
playwright==1.48.0
pytest==8.3.3
pytest-playwright==0.7.0
pyyaml==6.0
pytest-cov==6.0.0
coverage[toml]==7.6.9
allure-pytest==2.15.0
python-dotenv==1.0.1
